¿

+++ src/main/java/com/soft7/web/personSpace/dao/FrendsDAO.java
@@ -6,11 +6,11 @@ import java.sql.SQLException;
 import java.util.List;
 
 public interface FrendsDAO {
-    Long insertFriend(Friends friends) throws SQLException;
+    int insertFriend(Friends friends) throws SQLException;
 
-    int deleteFriendById(long id) throws SQLException;
+    int deleteFriendById(int id) throws SQLException;
 
-    Friends getFriendById(long id) throws SQLException;
+    Friends getFriendById(int id) throws SQLException;
 
     List<Friends> getAllFriends() throws SQLException;
 }
